Wrapping interval/timestamp handling up
